fix : User roles
유저 권한 개별테이블이 아닌 단독 테이블로 유지할 수 있도록 수정.
