+++ src/main/java/com/jaehyun/reservation/global/service/CustomUserDetailService.java
@@ -2,23 +2,25 @@ package com.jaehyun.reservation.global.service;
 
 import com.jaehyun.reservation.user.user.domain.entity.User;
 import com.jaehyun.reservation.user.user.domain.repository.UserRepository;
-import java.util.Optional;
-import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
-@RequiredArgsConstructor
 @Service
 public class CustomUserDetailService implements UserDetailsService {
 
-    private final UserRepository userRepository;
+  private final UserRepository userRepository;
 
-    public UserDetails loadUserByUsername(String loginId) throws UsernameNotFoundException {
-        return userRepository.findByLoginId(loginId)
-            .orElseThrow(() -> new UsernameNotFoundException("사용자를 찾을 수 없습니다."));
-    }
+  public CustomUserDetailService(UserRepository userRepository) {
+    this.userRepository = userRepository;
+  }
 
+  @Override
+  public UserDetails loadUserByUsername(String loginId) throws UsernameNotFoundException {
+    User user = userRepository.findByLoginId(loginId)
+        .orElseThrow(() -> new UsernameNotFoundException("사용자를 찾을 수 없습니다."));
 
+    return new JwtUserDetails(user);
+  }
 }
